multi_agents/requirements.txt
streamlit
opencv-python
pytesseract
python-dotenv
Pillow
numpy
openai